Clean up the custom paginator
