+++ nova/foundation/Providers/AppServiceProvider.php
@@ -6,11 +6,8 @@ use Inertia\Inertia;
 use Nova\Foundation\Nova;
 use Nova\Foundation\Macros;
 use Illuminate\Routing\Route;
-use Illuminate\Support\Collection;
-use Illuminate\Pagination\UrlWindow;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Blade;
-use Illuminate\Support\Facades\Request;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\View\Factory as ViewFactory;
@@ -84,89 +81,7 @@ class AppServiceProvider extends ServiceProvider
     protected function registerLengthAwarePaginator()
     {
         $this->app->bind(LengthAwarePaginator::class, function ($app, $values) {
-            return new class(...array_values($values)) extends LengthAwarePaginator {
-                public function only(...$attributes)
-                {
-                    return $this->transform(function ($item) use ($attributes) {
-                        return $item->only($attributes);
-                    });
-                }
-
-                public function transform($callback)
-                {
-                    $this->items->transform($callback);
-
-                    return $this;
-                }
-
-                public function toArray()
-                {
-                    return [
-                        'data' => $this->items->toArray(),
-                        'links' => $this->links(),
-                    ];
-                }
-
-                public function meta()
-                {
-                    return [
-                        'current_page' => $this->currentPage(),
-                        'first_page_url' => $this->url(1),
-                        'from' => $this->firstItem(),
-                        'last_page' => $this->lastPage(),
-                        'last_page_url' => $this->url($this->lastPage()),
-                        'next_page_url' => $this->nextPageUrl(),
-                        'path' => $this->path(),
-                        'per_page' => $this->perPage(),
-                        'prev_page_url' => $this->previousPageUrl(),
-                        'to' => $this->lastItem(),
-                        'total' => $this->total(),
-                    ];
-                }
-
-                public function links($view = null, $data = [])
-                {
-                    $this->appends(Request::all());
-
-                    $window = UrlWindow::make($this);
-
-                    $elements = array_filter([
-                        $window['first'],
-                        is_array($window['slider']) ? '...' : null,
-                        $window['slider'],
-                        is_array($window['last']) ? '...' : null,
-                        $window['last'],
-                    ]);
-
-                    return Collection::make($elements)->flatMap(function ($item) {
-                        if (is_array($item)) {
-                            return Collection::make($item)->map(function ($url, $page) {
-                                return [
-                                    'url' => $url,
-                                    'label' => $page,
-                                    'active' => $this->currentPage() === $page,
-                                ];
-                            });
-                        } else {
-                            return [
-                                [
-                                    'url' => null,
-                                    'label' => '...',
-                                    'active' => false,
-                                ],
-                            ];
-                        }
-                    })->prepend([
-                        'url' => $this->previousPageUrl(),
-                        'label' => 'Previous',
-                        'active' => false,
-                    ])->push([
-                        'url' => $this->nextPageUrl(),
-                        'label' => 'Next',
-                        'active' => false,
-                    ]);
-                }
-            };
+            return new \Nova\Foundation\LengthAwarePaginator(...array_values($values));
         });
     }
 }
